Rearrange further
I didn't like calling into process_handling module, then back into
things in the root then back into process_handling. So a bit of a tidy
up.

More tidying up will be needed at some point, however I see a lot of
this changing as I add in the ability to switch between engines so I'm
okay with leaving it in this state
